Adding file check in tests

+++ sdk/monitor/opentelemetry-exporter-azuremonitor/tests/trace/test_trace.py
@@ -27,12 +27,14 @@ STORAGE_PATH = os.path.join(TEST_FOLDER)
 
 # pylint: disable=invalid-name
 def setUpModule():
-    os.makedirs(TEST_FOLDER)
+    if not os.path.exists(TEST_FOLDER):
+        os.makedirs(TEST_FOLDER)
 
 
 # pylint: disable=invalid-name
 def tearDownModule():
-    shutil.rmtree(TEST_FOLDER)
+    if os.path.exists(TEST_FOLDER):
+        shutil.rmtree(TEST_FOLDER)
 
 
 def throw(exc_type, *args, **kwargs):
